+++ tests/fixtures/projects/modules/xs.module.kts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 module {
+  exclude()
   name("Math")
   members {
     add("Modules/Math/*.xs")